fix: update layout breaks
This change modifies the Layout component to conditionally include hr and
br tags based on the page type. If the page is not the home page, both hr
and br tags are added. Otherwise, only a br tag is added. This improves
the visual separation between the page metadata and the content.
The change also updates the DocsStepper component to include the
current page in the child pages, addressing the issue where some pages
were not included in the docsIndex.